notifications/requirements.txt
celery==5.2.7
Django==3.2.18
django-celery-results==2.4.0
djangorestframework==3.14.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.5
pyTelegramBotAPI==4.10.0
pytz==2022.6
redis==4.3.4
requests==2.28.2
six==1.16.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.3
urllib3==1.26.14